requirements.txt
python-dotenv>=1.0.0
langchain-core>=0.3.78,<1.0.0
langchain>=0.3.9
langchain_openai>=0.2.11
langchain_anthropic>=0.3.10
langchain_groq>=0.0.3
langchain_google_genai>=2.1.3
langchain_google_vertexai>=2.0.0,<3.0.0
groq>=0.4.2
openai>=1.6.0
anthropic>=0.49.0
e2b-code-interpreter==1.2.0b1
tenacity<9.0.0
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.6.1
langgraph==0.3.22
langgraph-checkpoint==2.0.23
tavily-python>=0.5.0
valyu>=0.1.0
requests>=2.32.0
typing-extensions>=4.0.0
pydantic>=2.5.2
tiktoken>=0.5.1
fastapi>=0.110.0
uvicorn>=0.28.0
starlette>=0.36.0
langsmith
langchain-mcp-adapters>=0.0.6
mcp>=1.4.1
numpy<2.0.0
plotly>=5.0.0
sse-starlette>=2.2.1
python-multipart>=0.0.6
nest-asyncio>=1.5.0

# File processing dependencies
PyPDF2>=3.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.2
Pillow>=10.0.1
pytesseract>=0.3.10
opencv-python>=4.8.1.78
python-magic>=0.4.27

# Audio/Video processing
SpeechRecognition>=3.10.0
moviepy>=1.0.3
whisper>=1.1.10

# Additional utilities for file handling
aiofiles>=23.2.1

# Slack integration
slack-sdk>=3.23.0

# PDF generation
reportlab>=4.0.0
